+++ src/main/java/subway/domain/application/LineService.java
@@ -1,10 +1,10 @@
-package subway.application;
+package subway.domain.application;
 
 import org.springframework.stereotype.Service;
-import subway.dao.LineDao;
+import subway.persistence.dao.LineDao;
 import subway.domain.Line;
-import subway.dto.LineRequest;
-import subway.dto.LineResponse;
+import subway.web.dto.LineRequest;
+import subway.web.dto.LineResponse;
 
 import java.util.List;
 import java.util.stream.Collectors;
