+++ src/main/java/subway/persistence/dao/StationDao.java
@@ -1,4 +1,4 @@
-package subway.dao;
+package subway.persistence.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
