+++ src/main/java/subway/web/dto/LineRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.web.dto;
 
 public class LineRequest {
     private String name;
