+++ src/main/java/subway/web/dto/LineResponse.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.web.dto;
 
 import subway.domain.Line;
