+++ src/main/java/subway/web/dto/StationRequest.java
@@ -1,4 +1,4 @@
-package subway.dto;
+package subway.web.dto;
 
 public class StationRequest {
     private String name;
