+++ src/main/java/subway/web/ui/StationController.java
@@ -1,10 +1,10 @@
-package subway.ui;
+package subway.web.ui;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import subway.dto.StationRequest;
-import subway.dto.StationResponse;
-import subway.application.StationService;
+import subway.web.dto.StationRequest;
+import subway.web.dto.StationResponse;
+import subway.domain.application.StationService;
 
 import java.net.URI;
 import java.sql.SQLException;
