+++ src/test/java/subway/integration/LineIntegrationTest.java
@@ -8,8 +8,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
-import subway.dto.LineRequest;
-import subway.dto.LineResponse;
+import subway.web.dto.LineRequest;
+import subway.web.dto.LineResponse;
 
 import java.util.HashMap;
 import java.util.List;
